changes in Categories classes

+++ src/main/java/com/example/demo/onlineshop/products/Products.java
@@ -49,7 +49,7 @@ public class Products {
         this.description = description;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
